шаблоны перенесены в отдельные файлы
