+++ src/components/sort.js
@@ -1,14 +1,5 @@
 import {createElement} from '../utils.js';
-
-const createSortTemplate = () => {
-  return (
-    `<div class="board__filter-list">
-      <a href="#" class="board__filter">SORT BY DEFAULT</a>
-      <a href="#" class="board__filter">SORT BY DATE up</a>
-      <a href="#" class="board__filter">SORT BY DATE down</a>
-    </div>`
-  );
-};
+import {createSortTemplate} from '../templates/sort.js';
 
 class Sort {
   constructor() {
