+++ src/components/tasks.js
@@ -1,10 +1,5 @@
 import {createElement} from '../utils.js';
-
-const createTasksTemplate = () => {
-  return (
-    `<div class="board__tasks"></div>`
-  );
-};
+import {createTasksTemplate} from '../templates/tasks.js';
 
 class Tasks {
   constructor() {
