+++ src/templates/board.js
@@ -1,0 +1,7 @@
+const createBoardTemplate = () => {
+  return (
+    `<section class="board container"></section>`
+  );
+};
+
+export {createBoardTemplate};
