+++ src/templates/loadMoreBtn.js
@@ -1,0 +1,7 @@
+const createLoadMoreBtnTemplate = () => {
+  return (
+    `<button class="load-more" type="button">load more</button>`
+  );
+};
+
+export {createLoadMoreBtnTemplate};
