+++ src/templates/tasks.js
@@ -1,0 +1,7 @@
+const createTasksTemplate = () => {
+  return (
+    `<div class="board__tasks"></div>`
+  );
+};
+
+export {createTasksTemplate};
